improve wishlist test
